Loaders::Association -> AssociationLoader
Moving the association loader to the HQ::GraphQL namespace
