--- lib/hq/graphql/loaders.rb
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-require "hq/graphql/loaders/association"
