+++ lib/hq/graphql/version.rb
@@ -2,6 +2,6 @@
 
 module HQ
   module GraphQL
-    VERSION = "2.0.10"
+    VERSION = "2.0.11"
   end
 end
